- OutputLayer, dC/dy, dC/dz 단위로 나누어서 계산하도록 수정 (이전 한번에 dC/dz를 구함) - SoftmaxLayer, cost()내에서 dC/dy만 구하고 나머지는 _backpropagation() 로직에 위임

+++ NeuralNetworksAndDeepLearning/src/layer/ConvLayer.h
@@ -148,8 +148,8 @@ protected:
 	virtual void _clearShape();
 	//virtual double _sumSquareGrad();
 	//virtual double _sumSquareParam();
-	virtual float sumSquareParamsData();
-	virtual float sumSquareParamsGrad();
+	virtual double sumSquareParamsData();
+	virtual double sumSquareParamsGrad();
 	virtual void _save(ofstream &ofs);
 	virtual void _load(ifstream &ifs, map<Layer *, Layer *> &layerMap);
 	virtual void update();
