+++ NeuralNetworksAndDeepLearning/src/layer/FullyConnectedLayer.h
@@ -138,8 +138,8 @@ protected:
 	virtual void _clearShape();
 	//virtual double _sumSquareGrad();
 	//virtual double _sumSquareParam();
-	virtual float sumSquareParamsData();
-	virtual float sumSquareParamsGrad();
+	virtual double sumSquareParamsData();
+	virtual double sumSquareParamsGrad();
 	virtual void scaleParamsGrad(DATATYPE scale);
 	virtual void _save(ofstream &ofs);
 	virtual void _load(ifstream &ifs, map<Layer *, Layer *> &layerMap);
